Termine generador, hice un generador de texto, imagen y audio

+++ 01-generador/scripts/hash.sh
@@ -1,15 +1,20 @@
 #!/bin/bash
 
-#Hashea el nombre del archivo pasado
-ARCHIVOS=$(ls outputs)
-OUTPUT_DIR=outputs
+#---
+#Recibe un directorio y un nombre de archivo perteneciente a ese directorio y
+#renombra el archivo con su hash correspondiente
+#---
 
+OUTPUT_DIR=$1
+FILENAME=$2
 
-for FILE in $ARCHIVOS
-do
-    md5sum $OUTPUT_DIR/$FILE > hash
-    mv $OUTPUT_DIR/$FILE $(cut -d " " -f 1 hash)
-    rm hash
-done
+if [[ ! -e $OUTPUT_DIR/$FILENAME ]]
+then
+    echo "Error! Archivo inexistente"
+    exit 1
+fi
 
-
+md5sum $OUTPUT_DIR/$FILENAME > hash
+mv $OUTPUT_DIR/$FILENAME $OUTPUT_DIR/$(cut -d " " -f 1 hash)
+rm hash
+exit 0
